decompress/requirements.txt
apache-beam[gcp]
google-cloud-storage
zstandard